Added manager/employee roles, refactored

+++ API/Controllers/TimestampsController.cs
@@ -23,19 +23,28 @@ namespace API.Controllers
         private readonly ITimestampRepository _timestampRepository;
         private readonly IUserRepository _userRepository;
         private readonly IJobsiteRepository _jobsiteRepository;
+        private readonly IUserAccessor _userAccessor;
 
         //constructor
-        public TimestampsController(IMapper imapper, ITimestampRepository timestampRepository, IUserRepository userRepository, IJobsiteRepository jobsiteRepository)
+        public TimestampsController(IMapper imapper, ITimestampRepository timestampRepository, 
+            IUserRepository userRepository, IJobsiteRepository jobsiteRepository,
+            IUserAccessor userAccessor)
         {
             _mapper = imapper;
             _timestampRepository = timestampRepository;
             _userRepository = userRepository;
             _jobsiteRepository = jobsiteRepository;
+            _userAccessor = userAccessor;
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] TimestampNewDto timestampNewDto)
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return this.StatusCode(StatusCodes.Status401Unauthorized, "Unauthorized: You must be a manager to perform this operation.");
+
             //get user
             var user = await _userRepository.GetUser(timestampNewDto.Username);
             if (user == null)
@@ -64,6 +73,10 @@ namespace API.Controllers
         [HttpDelete("{timestampId}")]
         public async Task<IActionResult> Delete(int timestampId)
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return this.StatusCode(StatusCodes.Status401Unauthorized, "Unauthorized: You must be a manager to perform this operation.");
 
             if (await _timestampRepository.DeleteTimestamp(timestampId))
                 return Ok("timestamp deleted");
@@ -76,6 +89,12 @@ namespace API.Controllers
         {
             try
             {
+                //manager status
+                var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+                if (loggedInUser.Manager == false)
+                    return this.StatusCode(StatusCodes.Status401Unauthorized, "Unauthorized: You must be a manager to perform this operation.");
+
+                //find timestamp
                 var timestamp = await _timestampRepository.GetTimestamp(timestampId);
                 if (timestamp == null)
                     return NotFound(new RestError(HttpStatusCode.NotFound, new { Timestamp = $"Timestamp with id {timestampId} not found" }));
@@ -102,10 +121,109 @@ namespace API.Controllers
             return BadRequest();
         }
 
+        //Get all timestamps from a user - can sort by date
+        [HttpGet("{username}")]
+        public async Task<ActionResult<UserInfoWithHoursWorkedDto>> GetAllTimeStamps(
+            string username, [FromQuery] TimestampParameters timestampParameters)
+        {
+            try
+            {
+                //manager status
+                var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+                if (loggedInUser.Manager == false)
+                    return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
+                var user = await _userRepository.GetUser(username);
+
+                //if user not found
+                if (user == null)
+                    return NotFound($"Username {username} not found.");
+
+                //Get timestamps by date
+                var filteredTimestamps = await _timestampRepository.GetTimestampsForUserByDate(user, timestampParameters);
+                user.Timestamps = filteredTimestamps;
+
+                //Create MetaData
+                var metadata = new
+                {
+                    filteredTimestamps.TotalCount,
+                    filteredTimestamps.PageSize,
+                    filteredTimestamps.CurrentPage,
+                    filteredTimestamps.HasNext,
+                    filteredTimestamps.HasPrevious
+                };
+
+                //Add metadata to header
+                Response.Headers.Add("X-Pagination", JsonConvert.SerializeObject(metadata));
+
+                var userWithTimestamps = _mapper.Map<UserInfoWithHoursWorkedDto>(user);
+                return Ok(userWithTimestamps);
+            }
+            catch (Exception)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, "Server Error: Failed to retrieve user data");
+            }
+        }
+
+        //Get all timestamps for a user by jobsite - can sort by date
+        [HttpGet("{username}/{moniker}")]
+        public async Task<ActionResult<UserInfoWithHoursWorkedDto>> GetAllTimestampsForUserByJobsite(string username,
+            string moniker, [FromQuery] TimestampParameters timestampParameters)
+        {
+            try
+            {
+                //manager status
+                var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+                if (loggedInUser.Manager == false)
+                    return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
+                var user = await _userRepository.GetUser(username);
+
+                //error if user not found
+                if (user == null)
+                    return NotFound($"Error: user '{username}' not found");
+
+                //error if jobsite not found
+                var jobsiteId = await _jobsiteRepository.GetJobsiteIdByMoniker(moniker);
+                if (jobsiteId == 0)
+                    return NotFound($"Error: jobsite '{moniker}' not found");
+
+                //filter user's timestamps by jobsite & date
+                var filteredTimestamps = await _timestampRepository
+                    .GetTimestampsForJobByUser(user, moniker, timestampParameters);
+
+                //add the timestamps to the user object
+                user.Timestamps = filteredTimestamps;
+
+                //Create MetaData
+                var metadata = new
+                {
+                    filteredTimestamps.TotalCount,
+                    filteredTimestamps.PageSize,
+                    filteredTimestamps.CurrentPage,
+                    filteredTimestamps.HasNext,
+                    filteredTimestamps.HasPrevious
+                };
+
+                //Add metadata to header
+                Response.Headers.Add("X-Pagination", JsonConvert.SerializeObject(metadata));
+
+                return Ok(_mapper.Map<UserInfoWithHoursWorkedDto>(user));
+            }
+            catch (Exception)
+            {
+                return this.StatusCode(StatusCodes.Status500InternalServerError, "Server Error: Failed to query database.");
+            }
+        }
 
         [HttpGet("info")]
         public async Task<ActionResult<object>> TimestampInfo()
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return this.StatusCode(StatusCodes.Status401Unauthorized, "Unauthorized: You must be a manager to perform this operation.");
+
             var timestamps = await _timestampRepository.GetTimestamps();
             var employeeCount = TimestampActions.UniqueEmployeeCount(timestamps);
             var jobsitesCount = TimestampActions.UniqueJobsiteCount(timestamps);
@@ -124,6 +242,11 @@ namespace API.Controllers
         public async Task<IActionResult> GetTimestamps(
             [FromQuery] TimestampParameters timestampParameters)
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
             //returns a paged list
             var timestamps = await _timestampRepository.GetTimestamps(timestampParameters);
 
@@ -148,6 +271,11 @@ namespace API.Controllers
         {
             try
             {
+                //manager status
+                var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+                if (loggedInUser.Manager == false)
+                    return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
                 var user = await _userRepository.GetUser(username);
 
                 //if user not found
@@ -178,6 +306,11 @@ namespace API.Controllers
         public async Task<IActionResult> GetWorkHistory(
             [FromQuery] WorkHistoryParameters workHistoryParameters)
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
             //max 45 days
             var timestamps = await _timestampRepository.GetTimestamps(workHistoryParameters);
             //form custom DTO based on timestamps
@@ -190,6 +323,11 @@ namespace API.Controllers
         public async Task<IActionResult> GetJobsitesVisited(
             [FromQuery] TimestampParameters timestampParameters)
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
             var timestamps = await _timestampRepository.GetTimestampsUnpaged(timestampParameters);
 
             //get Jobsites visited
@@ -219,13 +357,22 @@ namespace API.Controllers
         [HttpGet("clockedin")]
         public async Task<ActionResult<object>> CurrentlyClockedIn()
         {
+            //manager status
+            var loggedInUser = await _userRepository.GetUser(_userAccessor.GetCurrentUsername());
+            if (loggedInUser.Manager == false)
+                return Unauthorized(new RestError(HttpStatusCode.Unauthorized, new { Unauthorized = "Unauthorized to perform action" }));
+
             var timestamps = await _timestampRepository.TimestampsCurrentlyClockedIn();
             //get list of strings of employee names 
             var clockedInEmployees = TimestampActions.ClockedInEmployees(timestamps);
 
+            //get list of Clocked In Jobsites
+            var clockedInJobsites = TimestampActions.ClockedInJobsites(timestamps);
+
             var dto = new
             {
                 currentlyClockedIn = clockedInEmployees,
+                clockedInJobsites,
                 timestamps = _mapper.Map<ICollection<TimestampClockedInDto>>(timestamps)
             };
             return Ok(dto);
